фикс фишспич громкости

+++ src/voice_models/fish_speech_model.py
@@ -757,7 +757,7 @@ class FishSpeechModel(IVoiceModel):
                 return None
 
             stereo_output_path = raw_output_path.replace("_raw", "_stereo")
-            converted_file = self.parent.convert_wav_to_stereo(raw_output_path, stereo_output_path, volume=str(0.5+int(vol)))
+            converted_file = self.parent.convert_wav_to_stereo(raw_output_path, stereo_output_path, volume=str(0.5+float(vol)))
             
             processed_output_path = stereo_output_path if converted_file and os.path.exists(converted_file) else raw_output_path
             if processed_output_path == stereo_output_path:
